requirements.txt
Django==4.2.7
Pillow==10.0.1
stripe==7.8.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
python-decouple==3.8
django-allauth==0.57.0 